api/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
astroid==2.15.8
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
black==23.9.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
dnspython==2.4.2
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
parameterized==0.9.0
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
PyJWT==2.8.0
pylint==2.17.6
pymongo==4.5.0
pyparsing==3.1.1
pytest==7.4.2
pytest-asyncio==0.21.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
tomlkit==0.12.1
uritemplate==4.1.1
urllib3==2.0.5
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.2
